refactor(table): rename folder to table
