+++ packages/aksara-ui-core/src/components/index.ts
@@ -3,7 +3,6 @@ export * from './badge';
 export * from './blankslate';
 export * from './button';
 export * from './collapsible';
-export * from './data-table';
 export * from './dialog';
 export * from './dropdown';
 export * from './form';
@@ -16,5 +15,6 @@ export * from './panel';
 export * from './popover';
 export * from './side-sheet';
 export * from './skeleton';
+export * from './table';
 export * from './tabs';
 export * from './tooltip';
