+++ packages/aksara-ui-core/src/components/table/components/Table/index.ts
