+++ packages/aksara-ui-core/src/components/table/components/TableBody/index.ts
