+++ packages/aksara-ui-core/src/components/table/components/TableHead/TableHeadCell.tsx
