+++ packages/aksara-ui-core/src/components/table/index.ts
