+++ packages/aksara-ui-core/src/components/table/styles.ts
